requirements.txt
# Core
streamlit==1.37.1
python-dotenv==1.0.1

# LangChain + OpenAI
langchain==0.2.17
langchain-core==0.2.43
langchain-community==0.2.11
langchain-openai==0.1.14

# Vector Database
chromadb==0.4.24
tiktoken==0.7.0

# PDF Parsing & Generation
pypdf==3.17.4
reportlab==4.2.5

# Email + Network
requests==2.31.0

# Utilities
tenacity==8.5.0
packaging==24.2
